+++ src/fides/api/models/privacy_experience.py
@@ -48,6 +48,20 @@ class Layer1ButtonOption(Enum):
     OPT_IN_OPT_OUT = "opt_in_opt_out"
 
 
+class RejectAllMechanism(Enum):
+    """
+    Reject all mechanism options - not formalized in the db.
+    Used to configure the behavior of the reject all button in TCF experiences
+    """
+
+    # Reject both consent and legitimate interest preferences (all purposes, special features, vendors)
+    # This is the default behavior
+    REJECT_ALL = "reject_all"
+    # Reject only consent preferences (all purposes, special features, vendors).
+    # Do not reject any legitimate interest preferences.
+    REJECT_CONSENT_ONLY = "reject_consent_only"
+
+
 # Fides JS UX Types - there should only be one of these defined per region
 FidesJSUXTypes: List[ComponentType] = [
     ComponentType.banner_and_modal,
@@ -179,6 +193,10 @@ class PrivacyExperienceConfig(PrivacyExperienceConfigBase, Base):
     The Privacy Experience Configuration model that stores shared configuration for Privacy Experiences.
 
     - Translations, Notices, and Regions (via Privacy Experiences) are linked to this resource.
+
+    If you're adding a new PrivacyExperienceConfig, make sure to use the `create` method since it has
+    custom logic that ensures other resources are created/updated as needed, as well as setting the
+    expected values for some fields in the experience config itself.
     """
 
     allow_language_selection = Column(
@@ -202,6 +220,13 @@ class PrivacyExperienceConfig(PrivacyExperienceConfigBase, Base):
     origin = Column(
         String, ForeignKey(ExperienceConfigTemplate.id_field_path)
     )  # The template from which this config was created if applicable
+
+    # Mechanism to use when the reject all button is clicked in a TCF experience
+    # Nullable because this is not applicable for other experience types
+    reject_all_mechanism = Column(
+        EnumColumn(RejectAllMechanism),
+        nullable=True,
+    )
 
     # Relationships
     experiences = relationship(
@@ -303,6 +328,15 @@ class PrivacyExperienceConfig(PrivacyExperienceConfigBase, Base):
         )
         # Link Properties to this Privacy Experience config via the PrivacyExperienceConfigProperty table
         link_properties_to_experience_config(db, properties, experience_config)
+
+        # If the reject all mechanism is not set and the experience config is a TCF experience,
+        # set the reject all mechanism to REJECT_ALL
+        if (
+            experience_config.component == ComponentType.tcf_overlay
+            and experience_config.reject_all_mechanism is None
+        ):
+            experience_config.reject_all_mechanism = RejectAllMechanism.REJECT_ALL  # type: ignore
+            experience_config.save(db)
 
         return experience_config
 
@@ -506,6 +540,13 @@ class PrivacyExperienceConfigHistory(
         ForeignKey(ExperienceTranslation.id_field_path, ondelete="SET NULL"),
         index=True,
     )  # If a translation is deleted, this is set to null, but the overall record remains in the database for reporting purposes
+
+    # Mechanism to use when the reject all button is clicked in a TCF experience
+    # Nullable because this is not applicable for other experience types
+    reject_all_mechanism = Column(
+        EnumColumn(RejectAllMechanism),
+        nullable=True,
+    )
 
     version = Column(Float, nullable=False, default=1.0)